Update Id - from Int to Long + naming convention for Database items

+++ app/src/main/java/com/bandit/mapper/ConcertMapper.kt
@@ -14,14 +14,14 @@ object ConcertMapper : Mapper<Concert, ConcertDBEntry> {
             entry.country,
             entry.place,
             mapIntToConcertType(entry.type.toInt()),
-            entry.id.toInt(),
+            entry.id,
             entry.userUid
         )
     }
 
     override fun fromItemToDbEntry(item: Concert): ConcertDBEntry {
         return ConcertDBEntry(
-            item.id.toLong(),
+            item.id,
             item.name,
             item.dateTime.toString(),
             item.city,
